+++ src/main/java/ar/edu/itba/sds_2021_q1_g02/models/Position.java
@@ -1,26 +1,25 @@
 package ar.edu.itba.sds_2021_q1_g02.models;
 
-import java.math.BigDecimal;
 
 public class Position {
-    private BigDecimal x;
-    private BigDecimal y;
+    private double x;
+    private double y;
 
-    public Position(BigDecimal x, BigDecimal y) {
+    public Position(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public BigDecimal getX() {
+    public double getX() {
         return this.x;
     }
-    public void setX(BigDecimal x) { this.x  = x;}
+    public void setX(double x) { this.x  = x;}
 
-    public BigDecimal getY() {
+    public double getY() {
         return this.y;
     }
 
-    public void setY(BigDecimal y) {
+    public void setY(double y) {
         this.y = y;
     }
