+++ src/main/java/ar/edu/itba/sds_2021_q1_g02/models/Step.java
@@ -1,17 +1,16 @@
 package ar.edu.itba.sds_2021_q1_g02.models;
 
-import java.math.BigDecimal;
 import java.util.Map;
 
 public class Step {
     private final Map<Integer, Particle> particleMap;
     private final Map<Particle, Position> previousParticlesPosition;
     private final Map<Particle, Velocity> previousParticlesVelocity;
-    private final BigDecimal deltaTime;
-    private final BigDecimal absoluteTime;
+    private final double deltaTime;
+    private final double absoluteTime;
     private final int step;
 
-    public Step(Map<Integer, Particle> particleMap, Map<Particle, Position> previousParticlesPosition, Map<Particle, Velocity> previousParticlesVelocity, BigDecimal deltaTime, BigDecimal absoluteTime, int step) {
+    public Step(Map<Integer, Particle> particleMap, Map<Particle, Position> previousParticlesPosition, Map<Particle, Velocity> previousParticlesVelocity, double deltaTime, double absoluteTime, int step) {
         this.particleMap = particleMap;
         this.previousParticlesPosition = previousParticlesPosition;
         this.previousParticlesVelocity = previousParticlesVelocity;
@@ -21,11 +20,11 @@ public class Step {
         this.step = step;
     }
 
-    public BigDecimal getRelativeTime() {
+    public double getRelativeTime() {
         return this.deltaTime;
     }
 
-    public BigDecimal getAbsoluteTime() {
+    public double getAbsoluteTime() {
         return this.absoluteTime;
     }
